// lib/types/form-schema/property-owners-form-schema.ts
import {z} from "zod";

export const PropertyOwnerFormSchema = z.object({
    name: z.string().optional(),
    email: z.string().optional(),
    phone: z.string().optional(),
    address: z.string().optional(),
    mobile: z.string().optional(),
    resident_country: z.string().optional(),
    secondary_mobile: z.string().optional(),
    nationality: z.string().optional(),
    secondary_phone: z.string().optional(),
    area_owned: z.string().optional(),
    plot_number: z.string().optional(),
    total_area: z.string().optional(),
    usage_type: z.string().optional(),
    project: z.string().optional(),
    master_project: z.string().optional(),
    municipality_number: z.string().optional(),
    procedure_number: z.string().optional(),
    fools: z.string().optional(),
    common_area: z.string().optional(),
    parking_area: z.string().optional(),
    balcony_area: z.string().optional(),
    unit_number: z.string().optional(),
    registration_number: z.string().optional(),
    building_name: z.string().optional(),
    size: z.string().optional(),
    room_type: z.string().optional(),
})